requirements.txt
opencv-python
numpy
tensorflow-gpu
msgpack-python
msgpack-numpy

